Refactor #175 헬스 체크 API 도메인 패키지에서 Share 패키지로 이동
